# docs/contracts/api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: L9 ENRICH — Domain Enrichment API
  version: "2.3.1"
  description: |
    Universal domain-aware entity enrichment. Layer 2 of the L9 three-layer
    intelligence stack. Serves Salesforce Apex callouts, Odoo async_executor,
    and any L9 constellation node via PacketEnvelope.

    ## Authentication
    All endpoints (except GET /api/v1/health) require an `X-API-Key` header.
    The raw key is SHA-256 hashed and compared against `API_KEY_HASH` using
    constant-time comparison (`hmac.compare_digest`).

    ## Rate Limiting
    Global: 120 requests/minute per client IP.
    Source: `app/middleware/rate_limiter.py`
  contact:
    name: L9 Enrichment Engine
    url: https://github.com/cryptoxdog/Enrichment.Inference.Engine
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://enrich.yourdomain.com
    description: Production
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        SHA-256-hashed API key. Callers send the raw key; the server hashes and
        compares. Source: app/core/auth.py
  schemas:
...
    ErrorResponse:
      $ref: "./schemas/error-responses.yaml#/components/schemas/ErrorResponse"

    # ── Enrichment ──────────────────────────────────────────
...
    DiscoverRequest:
      title: DiscoverRequest
      type: object
      required:
        - entity_id
        - domain
        - object_type
        - tenant_id
      properties:
        entity_id:
          type: string
        domain:
          type: string
        object_type:
          type: string
        tenant_id:
          type: string
      x-source-file: app/api/v1/discover.py:27
      additionalProperties: false
    ScanRequest:
      title: ScanRequest
      description: CRM field scan request for Seed tier entry point.
      type: object
      required:
        - fields
        - domain
        - tenant_id
      properties:
        fields:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              sample_values:
                type: array
                items: {}
                nullable: true
              fill_rate:
                type: number
                nullable: true
        domain:
          type: string
          example: plasticos
        tenant_id:
          type: string
          example: acme-corp
      x-source-file: app/api/v1/discover.py:37
      additionalProperties: false
    ApprovalRequest:
      title: ApprovalRequest
      type: object
      required:
        - approved
        - reviewed_by
      properties:
        approved:
          type: boolean
        reviewed_by:
          type: string
          description: Identity of the reviewer.
      x-source-file: app/api/v1/discover.py:47
      additionalProperties: false
  responses:
    Unauthorized:
      description: Missing X-API-Key header
      content:
        application/json:
          schema:
            $ref: "./schemas/error-responses.yaml#/components/schemas/ErrorResponse"
          example:
            detail: "Missing X-API-Key header"
...
    InternalError:
      description: >-
        Internal server error. The `detail` is a generic, non-sensitive message;
        the underlying exception is recorded in server-side structured logs only
        and is never returned to the client (CWE-209 information-exposure
        hardening applied to the discover, scan, converge, and attestation
        handlers).
      content:
        application/json:
          schema:
            $ref: "./schemas/error-responses.yaml#/components/schemas/ErrorResponse"
          example:
            detail: "Internal enrichment error"
paths:
...
  /api/v1/discover:
    post:
      operationId: discoverSchema
      summary: Trigger schema discovery for a domain entity
      description: |
        Runs schema discovery for an entity within a domain. Analyzes the entity's
        existing fields, gaps, and coverage against the domain YAML spec. Returns
        discovered field candidates and coverage metrics.
      tags: [discover]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoverRequest"
            example:
              entity_id: "acct_001"
              domain: plasticos
              object_type: Account
              tenant_id: acme-corp
      responses:
        "200":
          description: Schema discovery result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
      x-source-file: app/api/v1/discover.py:51
  /api/v1/scan:
    post:
      operationId: scanCRMFields
      summary: CRM field scan — Seed tier entry point
      description: |
        Scans CRM field metadata against the domain YAML spec. Returns coverage ratio,
        missing fields, enrichment candidates, and cost estimates. This is the primary
        "Seed tier" entry point for onboarding new CRM instances.
      tags: [discover]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
            example:
              fields:
                - name: company_name
                  type: string
                  sample_values: ["Acme Corp", "Plastics Inc"]
                  fill_rate: 0.98
                - name: industry
                  type: string
                  fill_rate: 0.72
              domain: plasticos
              tenant_id: acme-corp
      responses:
        "200":
          description: CRM field scan result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Domain not found — no matching domain spec in the registry
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
      x-source-file: app/api/v1/discover.py:72
  /api/v1/proposals/{domain}:
    get:
      operationId: getSchemaProposals
      summary: Get pending schema proposals for a domain
      tags: [discover]
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
          example: plasticos
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
          description: Tenant identifier.
          example: acme-corp
      responses:
        "200":
          description: List of pending schema proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    field_name:
                      type: string
                    field_type:
                      type: string
                    source:
                      type: string
                    fill_rate:
                      type: number
                    avg_confidence:
                      type: number
                    sample_values:
                      type: array
                    proposed_gate:
                      type: string
                      nullable: true
                    proposed_scoring_dimension:
                      type: string
                      nullable: true
                    yaml_diff:
                      type: string
                      nullable: true
                    approval_status:
                      type: string
                      enum: [pending, approved, rejected]
                    created_at:
                      type: string
                      format: date-time
      x-source-file: app/api/v1/discover.py:93
  /api/v1/proposals/{proposal_id}/approve:
    post:
      operationId: approveSchemaProposal
      summary: Approve or reject a schema proposal
      tags: [discover]
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequest"
            example:
              approved: true
              reviewed_by: "igor@example.com"
      responses:
        "200":
          description: Approval status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [approved, rejected]
                  proposal_id:
                    type: string
                    format: uuid
      x-source-file: app/api/v1/discover.py:113

  # ── Fields ─────────────────────────────────────────────────
...